api: created mock lang app to prevent calling real language
Tests that required a "compile" ended up calling an actual language, typically L1. This came with the latency of a cold start of L1. This change creates a mock language that allows the user to set the data that should be returned if the mock language is called with specific code.